Fixes type check

+++ src/Marketplace.php
@@ -592,7 +592,7 @@ class Marketplace extends BasePlugin
                     LogToFile::info(json_encode($lineItem), 'marketplace');
 
                     // If there isn’t a payee or a total on this line item, nothing to do
-                    if (!$payeeCurrent || $lineItem->total === 0) {
+                    if (!$payeeCurrent || $lineItem->total === (float) 0) {
                         break;
                     }
